yt6801: remove xz compression of driver
no need to compress since other drivers are
not compressed.

had one instance that when system boots up
from the flash image USB drive, and when
setup the network, the driver is not loaded.
it is caused when kernel is rebuilt, yt6801
also need to be rebuilt, but linux-initrd
initramfs is not rebuilt which actually
runs the module dependencies, so linux-initrd
also need to rebuilt together.

Signed-off-by: Vincent Li <vincent.mc.li@gmail.com>